+++ lib/plugin-event-handlers.ts
@@ -35,7 +35,9 @@ import { ensureIsAbsolute } from "./helpers/paths";
 
 import {
   createTimestamp,
+  orderMessages,
   removeDuplicatedStepDefinitions,
+  removeUnusedDefinitions,
 } from "./helpers/messages";
 
 import { memoize } from "./helpers/memoize";
@@ -310,7 +312,12 @@ export async function beforeRunHandler(config: Cypress.PluginConfigOptions) {
   };
 }
 
-export async function afterRunHandler(config: Cypress.PluginConfigOptions) {
+export async function afterRunHandler(
+  config: Cypress.PluginConfigOptions,
+  results:
+    | CypressCommandLine.CypressRunResult
+    | CypressCommandLine.CypressFailedRunResult,
+) {
   debug("afterRunHandler()");
 
   const preprocessor = await resolve(config, config.env, "/");
@@ -329,10 +336,7 @@ export async function afterRunHandler(config: Cypress.PluginConfigOptions) {
 
   const testRunFinished: messages.Envelope = {
     testRunFinished: {
-      /**
-       * We're missing a "success" attribute here, but cucumber-js doesn't output it, so I won't.
-       * Mostly because I don't want to look into the semantics of it right now.
-       */
+      success: "totalFailed" in results ? results.totalFailed === 0 : false,
       timestamp: createTimestamp(),
     } as messages.TestRunFinished,
   };
@@ -349,6 +353,7 @@ export async function afterRunHandler(config: Cypress.PluginConfigOptions) {
   };
 
   removeDuplicatedStepDefinitions(state.messages.accumulation);
+  removeUnusedDefinitions(state.messages.accumulation);
 
   if (preprocessor.messages.enabled) {
     const messagesPath = ensureIsAbsolute(
@@ -645,8 +650,8 @@ export async function afterSpecHandler(
         state: "after-spec",
         pretty: state.pretty,
         messages: {
-          accumulation: state.messages.accumulation.concat(
-            state.messages.current,
+          accumulation: orderMessages(
+            state.messages.accumulation.concat(state.messages.current),
           ),
         },
       };